restructuring and splitting of SagaStream code into several classes.
