+++ simplesaga-serialization/src/main/avro/ActionSchemas.avdl
@@ -73,6 +73,7 @@ protocol SagaSchemas {
     string actionStatus;
     union { array<AvroSagaError>, null } actionErrors;
     union { AvroActionUndoCommand, null } undoCommand;
+    boolean inUndo;
   }
 
   record AvroSagaTransitionSagaStatusChange {
